-加入了生成二维码的功能
