+++ Assets/Source/QRCode/EasyCodeScanner.cs
