+++ Assets/Source/QRCode/QRGenerateManger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7c23b7861dbaec4f930f56e8704e86b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
