+++ Assets/Source/QRCode/QRScanManager.cs
