+++ Assets/Source/QRCode/QRScanManager.cs.meta
